after pairing

+++ spec/jobs/foreign_key_polymorphic_association_job_spec.rb
@@ -14,13 +14,14 @@ describe ForeignKeyPolymorphicAssociationJob, :postgres do
   let!(:sil) { SpecialIssueList.create(appeal: appeal) }
   # let(:task) { create(:distribution_task) }
 
-  context "_id is nil" do
+  context "_id is nil regardless of existence of associated record" do
     before do
       sil.update_column(:appeal_id, nil)
-      appeal.destroy!
+      sil.update_column(:appeal_type, nil)
+      appeal.destroy! if [true, false].sample
     end
     it "does not send alert" do
-      expect(Appeal.count).to eq 0
+      expect(Appeal.count).to eq(0).or eq(1)
       subject
       expect(slack_service).not_to have_received(:send_notification)
     end
@@ -34,6 +35,19 @@ describe ForeignKeyPolymorphicAssociationJob, :postgres do
   end
 
   context "_id is nil but _type is non-nil" do
+    before do
+      sil.update_column(:appeal_id, nil)
+      appeal.destroy! if [true, false].sample
+    end
+    it "sends alert" do
+      expect(Appeal.count).to eq(0).or eq(1)
+      expect(sil.appeal_type).not_to eq nil
+
+      subject
+
+      message = "Found SpecialIssueList record with nil appeal_id but non-nil appeal_type: [#{sil.id}]"
+      expect(slack_service).to have_received(:send_notification).with(message).once
+    end
   end
 
   # This is the main objective of the job
@@ -73,6 +87,32 @@ describe ForeignKeyPolymorphicAssociationJob, :postgres do
         message = /Found .* orphaned record/
         expect(slack_service).to have_received(:send_notification).with(message).twice
       end
+    end
+  end
+
+  context "when checking Claimant.participant_id foreign key" do
+    let(:claimant) { appeal.claimant }
+    context "associated Person exists" do
+      it "does not send alert" do
+        expect(claimant.person).not_to eq nil
+        expect(Person.find_by_participant_id(claimant.participant_id)).not_to eq nil
+        subject
+        expect(slack_service).not_to have_received(:send_notification)
+      end
+    end
+    context "associated Person does not exist" do
+      before do
+        claimant.person.destroy!
+      end
+      # it "sends alert" do
+      #   expect(Person.find_by_participant_id(claimant.participant_id)).to eq nil
+      #   binding.pry
+
+      #   subject
+
+      #   message = "Found SpecialIssueList record with nil appeal_id but non-nil appeal_type: [#{sil.id}]"
+      #   expect(slack_service).to have_received(:send_notification).with(message).once
+      # end
     end
   end
 end
